Fix translator and finish TODOs
